Add "Interim finance committee role"
While the current finance committee is only "interim" and will hopefully
be replaced soon with a more permanent construct, this committee is
beyond the theoretical planning stage and is distributing real money.
That will have an impact on the the community one way or another.
In the interest of transparancy I think it is thus warrented at this point
to list the team members on the team page. At least that's where I would
look for it if I did not know whom to complain to ;-)

Given that it's "interim", I've placed the committee between the
"coordinators" and the "maintainers" on the page, but I suggest
revisiting that once the full Astropy governance structure is worked out.