Limit scope of affilaited pacakges
This PR adds a bullet point to clearify that affilaited pacakges should
be in some way related to other astropy packages or functionality,
not just be any pacakge that might be useful for astronomers.
(E.g., we don't want to make numpy an affilated package, nor django,
just because astronoemrs can also build websites).