Update registry.json
From the abstract of our paper (DOI : 10.1016/j.ascom.2021.100503) we cite:

BayesicFitting is a comprehensive, general-purpose toolbox for simple and standardized model fitting. Its fitting options range from simple least-squares methods, via maximum likelihood to fully Bayesian inference, working on a multitude of available models. BayesicFitting is open source and has been in development and use since the 1990s. It has been applied to a variety of science applications, chiefly in astronomy.

BayesicFitting consists of a collection of PYTHON classes that can be combined to solve quite complicated inference problems. Amongst the classes are models, fitters, priors, error distributions, engines, samples, and of course NestedSampler, our general-purpose implementation of the nested sampling algorithm.

Nested sampling is a novel way to perform Bayesian calculations. It can be applied to inference problems, that consist of a parameterized model to fit measured data to. NestedSampler calculates the Bayesian evidence as the numeric integral over the posterior probability of (hyper)parameters of the problem. The solution in terms of the parameters is obtained as a set of weighted samples drawn from the posterior.