Replace Moritz with Aarya on the finance committee
Moritz has said in this nomination for the Coco, that he would step back
from other Astropy roles if elected. Based on nominations from the community,
the Coco has contacted Aarya, who is interested in joining the finance
committee.
This PR will stay open for two weeks for any discussion.