Add Coco policies and communication to teams page
Following the policies and reading/writing the communication
is part of the duties of the members of this committe
so it seem appropriate to list that here.
Also, the astropy.org/teams page is the best location where
teams and resposibilites are described. It's
indexed by search engines ans relatively easy to find,
so it seem useful to add this info here instead of
creating a new page just for this.
I'm not worried about making the text
a little longer, since this is rearely read top-to-bottom.