Reorganize Learn team
The previous leadership team does not have the time to lead learn any more, but they are going to be involved as contributors going forward. We discussed this today at a telecon and decided to mark the leadership role as "unfilled".
Additionally, with the current level of effort and structure, the split between leanr infracstruture and editors is somewhat artificial, e.g. people listed on the infrastructure side might "merge and review PRs" if it's only about fixing things, although that was previously described as the responsibility of the editors only.

If the roles grow again, we can split this out again at a later time.