Add Derek and Moritz to website maintainers for sites.json
Updates to sites.json compromise the vast majority of all PRs to the website repro in the last few years and Derek and Moritz took care of most of those. So, in an effort to keep the list of people current with what everyone does, this PR adds Derek and Moritz as maintainers for the repro filling the role left by Adrian.

It also removes the "keep list of roles" role, since that's listed for the Coco already.