Add a role for security team
While experience in software security would be a great benefit for this role, it is not strictly necessary. Instead, the main focus of this role is to keep track of incoming reports, decide whom to contact about it (relevant maintainers, outside experts, feedback from institutional users, etc.) and to organize the process of how to respond to software security issues.
I suggest to also give this team the power to conduct or initiate security audits, but I envision that that will only happen if outside expertize is available through some collaboration or grant.