Add a new role for the maintainer mentorship program
We've wanted a maintainer mentorship program for a long time and, after many start,s it seems to get off the ground now. There are two active contributors that are currently paired with mentors. That's a start!
So, we suggest a new role to put this on some more permanent footing. It is important to list the role on the teams page, so that people know how to reach out and whom to contact.