Return a link to the description of badges
This change is the result of discussions as the Astropy 2025 coordination meeting and between the current Astropy affiliated package editors.

On the one hand, we want to provide users a quick way to find high-quality, integrated packages that are useful for their projects; on the other hand, we recognize that users have different needs. Some need packages compatible with cutting-edge versions, other need stable packages that don't require updating, e.g. teaching materials with every new release of astropy or numpy.
The text here tries to capture that in a few words as possible.