Remove packages that went through PyOpenSci from our old listing (#675)
* Remove packages that went through PyOpenSci from our old listing

Since the packges are now listed through PyOpenSci, they appear twice on the Astropy affiliated website once under the heading "Affiliated Packages Registry" and once under the heading "Affiliated Packages Registry (Pre-APE 22)".

We're trying to make that list the most useful for people to find packages and we think it's clearer if each package is listed only once. Thus, this PR removes the older (pre-APE 22) listing towards the bottom on the page.

Both affected packages have been contacted and agree:

https://github.com/StingraySoftware/stingray/issues/926
https://github.com/Cosmoglobe/zodipy/issues/50

* Remove extra blank line

Co-authored-by: Derek Homeier <dhomeie@gwdg.de>

---------

Co-authored-by: P. L. Lim <2090236+pllim@users.noreply.github.com>
Co-authored-by: Derek Homeier <dhomeie@gwdg.de>