Split out a new section Archived/emeritus for affiliated packages
The original intent (at least in my mind) for affiliated packages was to provide the community with a list of high-quality, actively maintained packages for their needs in astronomy. After many years, we have a large number of affiliated packages (good!) and some of them are no longer active, e.g. they have been archived in github by their maintainers. Yet, some of these packages are still used widely and perform well, at least if installed in environments with older Python versions. Thus, we don't want to remove them from the list, but we also want to distinguish them in some way.

Here, we split them out into a new section.

This was discussed at the 2025 Astropy Coordination meeting, but is a variant of of the ideas mentioned in the discussion.